requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
load-dotenv==0.1.0
python-dotenv==0.21.0
requests==2.28.1
telebot==0.0.4
urllib3==1.26.13
